# Conflicts:
#	HackEnrollment.c
#	cmake-build-debug/.ninja_deps
#	cmake-build-debug/.ninja_log
#	cmake-build-debug/Testing/Temporary/LastTest.log
#	cmake-build-debug/ex1_idanLifshitz_yuvalCarmeli.exe

+++ IsraeliQueue.c
@@ -374,7 +374,7 @@ IsraeliQueue IsraeliQueueMerge(IsraeliQueue* queue ,ComparisonFunction compFunc)
 
 int checkFriend(IsraeliQueue queue, void* item1, void* item2)
 {
-    int sum = 0, sumCount = 0, currItem;
+    double sum = 0, sumCount = 0, currItem;
     bool flag = false;
     for (int i = 0; queue->friendFunc[i] != NULL; i++) //for friend function in queue.
     {
